+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/wasbootstrap/configuration/RpConfiguration.java
@@ -44,6 +44,6 @@ public class RpConfiguration {
 
     public static class Sampling {
 
-        public double percentage = 100;
+        public float percentage = 100;
     }
 }
